Move permissions-related API tests
